Module cargo_update::ops

Main functions doing actual work.

Use installed_registry_packages() to list the installed packages, then use intersect_packages() to confirm which ones should be updated, poll the packages’ latest versions by calling RegistryPackage::pull_version() on them, continue with doing whatever you wish.

Structs

CargoConfig

cargo configuration, as obtained from .cargo/config[.toml]

GitRepoPackage

A representation of a package a remote git repository.

PackageConfig

Compilation configuration for one crate.

RegistryPackage

A representation of a package from the main crates.io repository.

Enums

ConfigOperation

A single operation to be executed upon configuration of a package.

PackageFilterElement

One of elements with which to filter required packages.

Functions

assert_index_path

Get the location of the registry index corresponding ot the given URL; if not present – make it and its parents.

cargo_hash

Stolen from and equivalent to short_hash() from https://github.com/rust-lang/cargo/blob/74f2b400d2be43da798f99f94957d359bc223988/src/cargo/util/hex.rs

crate_versions

Parse the raw crate descriptor from the repository into a collection of Semvers.

find_git_db_repo

Find the bare git repository in the specified directory for the specified crate

find_package_data

Find package data in the specified cargo index tree.

find_proxy

Check if there’s a proxy specified to be used.

get_index_url

Get the URL to update index from and the cargo name for it from the config file parallel to the specified crates file

installed_git_repo_packages

List the installed packages at the specified location that originate from a remote git repository.

installed_registry_packages

List the installed packages at the specified location that originate from the a cargo registry.

intersect_packages

Filter out the installed packages not specified to be updated and add the packages you specify to install, if they aren’t already installed via git.

registry_shortname

The short filesystem name for the repository, as used by cargo

resolve_crates_file

Follow install.root in the config file parallel to the specified crates file up to the final one.

update_index

Update the specified index repository from the specified URL.