Struct cargo_update::ops::MainRepoPackage

pub struct MainRepoPackage {
    pub name: String,
    pub version: Semver,
    pub newest_version: Option<Semver>,
}

A representation of a package from the main crates.io repository.

The newest version of a package is pulled from crates.io via pull_version().

The parse() function parses the format used in $HOME/.cargo/.crates.toml.

Examples

let package_s = "racer 1.2.10 (registry+https://github.com/rust-lang/crates.io-index)";
let mut package = MainRepoPackage::parse(package_s).unwrap();
assert_eq!(package,
           MainRepoPackage {
               name: "racer".to_string(),
               version: Semver::parse("1.2.10").unwrap(),
               newest_version: None,
           });

package.pull_version(crates_token);
assert!(package.newest_version.is_some());

Fields
name: String

The package's name.

Go to https://crates.io/crates/{name} to get the crate info.

version: Semver

The package's locally installed version.

newest_version: Option<Semver>

The latest version of the package vailable at the main crates.io repository.

None by default, acquire via MainRepoPackage::pull_version().

Methods

impl MainRepoPackage

fn parse(what: &str) -> Option<MainRepoPackage>

Try to decypher a package descriptor into a MainRepoPackage.

Will return None if:

the given package descriptor is invalid, or

the package descriptor is not from the main crates.io registry.

In the returned instance, newest_version is always None, get it via MainRepoPackage::pull_version().

Examples

Main repository packages:

let package_s = "racer 1.2.10 (registry+https://github.com/rust-lang/crates.io-index)";
assert_eq!(MainRepoPackage::parse(package_s).unwrap(),
           MainRepoPackage {
               name: "racer".to_string(),
               version: Semver::parse("1.2.10").unwrap(),
               newest_version: None,
           });

let package_s = "cargo-outdated 0.2.0 (registry+https://github.com/rust-lang/crates.io-index)";
assert_eq!(MainRepoPackage::parse(package_s).unwrap(),
           MainRepoPackage {
               name: "cargo-outdated".to_string(),
               version: Semver::parse("0.2.0").unwrap(),
               newest_version: None,
           });

Git repository:

let package_s = "treesize 0.2.1 (git+https://github.com/melak47/treesize-rs#v0.2.1)";
assert!(MainRepoPackage::parse(package_s).is_none());

fn pull_version(&mut self, crates_token: &str)

Download the version list for this crate off the main crates.io registry.

The provided token might or might not need to be valid, investigation ongoing.

Examples

let package_s = "racer 1.2.10 (registry+https://github.com/rust-lang/crates.io-index)";
let mut package = MainRepoPackage::parse(package_s).unwrap();
package.pull_version(crates_token);
assert!(package.newest_version.is_some());