Module cargo_update::ops

Main functions doing actual work.

Use installed_main_repo_packages() to list the installed packages, then use intersect_packages() to confirm which ones should be updated, poll the packages' latest versions by calling MainRepoPackage::pull_version on them, continue with doing whatever you wish.

Structs

MainRepoPackage
A representation of a package from the main crates.io repository.

Functions

crate_versions
Parse the raw crate descriptor from the repository into a collection of Semvers.

find_package_data
Find a package in the cargo index.

get_index_path
Get the location of the latest registry index in the specified cargo directory.

installed_main_repo_packages
List the installed packages at the specified location that originate from the main crates.io registry.

intersect_packages
Filter out the installed packages not specified to be updated.