Module cargo_update::ops

Main functions doing actual work.

Use installed_main_repo_packages() to list the installed packages, then use intersect_packages() to confirm which ones should be updated, poll the packages' latest versions by calling MainRepoPackage::pull_version() on them, continue with doing whatever you wish.

Structs

GitRepoPackage
A representation of a package a remote git repository.

MainRepoPackage
A representation of a package from the main crates.io repository.

PackageConfig
Compilation configuration for one crate.

Enums

ConfigOperation
A single operation to be executed upon configuration of a package.

PackageFilterElement
One of elements with which to filter required packages.

Functions

crate_versions
Parse the raw crate descriptor from the repository into a collection of Semvers.

find_git_db_repo
Find the bare git repository in the specified directory for the specified crate

find_package_data
Find package data in the specified cargo index tree.

find_proxy
Check if there's a proxy specified to be used.

get_index_path
Get the location of the latest registry index whose name optionally starts with the registry URL's domain name in the specified cargo directory.

get_index_url
Get the URL to update index from the config file parallel to the specified crates file

installed_git_repo_packages
List the installed packages at the specified location that originate from a remote git repository.

installed_main_repo_packages
List the installed packages at the specified location that originate from the main crates.io registry.

intersect_packages
Filter out the installed packages not specified to be updated and add the packages you specify to install, if they aren't already installed via git.

resolve_crates_file
Follow install.root in the config file parallel to the specified crates file up to the final one.

update_index
Update the specified index repository from the specified URL.